requirements.txt
# Discord
discord.py[voice]==2.3.2

# Music
wavelink==3.4.1

# Utils
python-dotenv==1.0.0
aiohttp==3.9.1

# Audio
PyNaCl==1.5.0
